pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Resume Analysis and Question Generation Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "python-multipart",
    "uvicorn",
    "google-generativeai",
    "python-docx",
    "PyPDF2",
    "python-dotenv",
    "pdfplumber",
]
